initial commit i would normally have not done until this last one was completed. i was looking to include something prior to syncing up with the recruitment team hence the 2 commits.
validation, rest calls are functional, documentation, etc.
httpresponse handling has been left raw with no custom exception handling and just letting whatever throws bubble up as not sure if this is intended to be a client facing api or an internal development facing api to give more details to the client developer.
in the case of a public api this should be managed more appropriately and not necessarily have the inner workings exposed for security/vulnerability reasons.
quite possible that some of the iterations could be cleaner with more in depth knowledge on current ef capabilities however this seemed practical for the exercise.